% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{DecayChunk}
\alias{DecayChunk}
\title{DecayChunk()}
\usage{
DecayChunk(wm_Box, time_current, chunk_Lifecyle)
}
\arguments{
\item{wm_Box}{A data frame for working memory storage}

\item{time_current}{Current time}

\item{chunk_Lifecyle}{A data frame with all chunk's birth and death}
}
\value{
A list of chunk lifecycle and working memory
}
\description{
Make a chunk be eliminated
}
\examples{
DecayChunk(wm_Box, time_current, chunk_Lifecyle)
}
