% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initiation_data.R
\name{DefineScenario}
\alias{DefineScenario}
\title{Set scenario}
\usage{
DefineScenario(user_scn)
}
\arguments{
\item{user_scn}{If a user send a scenario (CSV), then add it into the model's inner variable}
}
\value{
task_1 The scenario file that will be used in the model
}
\description{
Set scenario
}
\examples{
DefineScenario(csv)
}
