% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{ExtChunk}
\alias{ExtChunk}
\title{ExtChunk()}
\usage{
ExtChunk(
  line_of_input,
  oper,
  time_current,
  time_past,
  wm_Box,
  oper_Time,
  num_of_chunks,
  chunk_Lifecyle
)
}
\arguments{
\item{line_of_input}{Line number}

\item{oper}{Operator name}

\item{time_current}{Current time}

\item{time_past}{Time when a previous operator was added}

\item{wm_Box}{A data frame for working memory storage}

\item{oper_Time}{Operator time}

\item{num_of_chunks}{Number of chunk in one line}

\item{chunk_Lifecyle}{A data frame with all chunk's birth and death}
}
\value{
A list of chunk lifecycle and working memory
}
\description{
Extracting and pushing a chunk (Main function)
}
\examples{
ExtChunk(line_of_input, oper, time_current, time_past, wm_Box, oper_Time, num_of_chunks, chunk_Lifecyle)
ExtChunk(3, "Look", 5000, 4200, working memory, 550, 1, chunk_Lifecycle)
}
