% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCT.R
\name{ExtOper}
\alias{ExtOper}
\title{ExtOper()}
\usage{
ExtOper(line_of_input)
}
\arguments{
\item{line_of_input}{}
}
\value{
operator name
}
\description{
Return a pure operator name
}
\examples{
ExtOper(line_of_input)
}
