% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{GetActivation}
\alias{GetActivation}
\title{GetActivation()}
\usage{
GetActivation(stack_Depth, last_Time, rehearsals)
}
\arguments{
\item{stack_Depth}{The floor number of a chunk in working memory storage (wm_Box)}

\item{last_Time}{Current time}

\item{rehearsals}{Number of rehearsals}
}
\value{
Calculated activation value
}
\description{
Calculation for activation
}
\examples{
GetActivation(stack_Depth, last_Time, rehearsals)
}
