% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{GetLOI}
\alias{GetLOI}
\title{GetLOI()}
\usage{
GetLOI(op_Name, err_chunk_made)
}
\arguments{
\item{op_Name}{Operator name}

\item{err_chunk_made}{Error chunk}
}
\value{
Line
}
\description{
Line of input generator
}
\examples{
GetLOI(op_Name, err_chunk_made)
}
