% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{GetMemoryChunk}
\alias{GetMemoryChunk}
\title{GetMemoryChunk()}
\usage{
GetMemoryChunk(chunk_Lifecyle, acc_Time_1)
}
\arguments{
\item{chunk_Lifecyle}{A table with all chunk's birth and death}

\item{acc_Time_1}{Entire task duration}
}
\value{
A list of chunk lifecyle and memory load
}
\description{
Calculate and return memory load
}
\examples{
GetMemoryChunk(chunk lifecycle, total task completion time)
}
