% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{GetNumOper}
\alias{GetNumOper}
\title{GetNumOper()}
\usage{
GetNumOper(oper_Name, num_Oper, oper_Time)
}
\arguments{
\item{oper_Name}{Operator name}

\item{num_Oper}{Number of operators}

\item{oper_Time}{Operator time (duration)}
}
\value{
Number of operator
}
\description{
Return the number of operators (Calculate repetition for novices)
}
\examples{
GetNumOper(oper_Name, num_Oper, oper_Time)
}
