% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{GetRehearsal}
\alias{GetRehearsal}
\title{GetRehearsal()}
\usage{
GetRehearsal(oper, chk_nam_list, current_rehearsal, same, chunk_Name)
}
\arguments{
\item{oper}{Operator name}

\item{chk_nam_list}{A list of chunk names in working memory storage}

\item{current_rehearsal}{The number of rehearsals}

\item{same}{A variable to check if the new chunk is the same with one of chunks in working memory storage}

\item{chunk_Name}{Chunk name}
}
\value{
Number of rehearsals
}
\description{
Calculate the number of rehearsals
}
\examples{
GetRehearsal(oper, chk_nam_list, current_rehearsal, same, chunk_Name)
}
