% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{GetStackDepth}
\alias{GetStackDepth}
\title{GetStackDepth()}
\usage{
GetStackDepth(chk_name)
}
\arguments{
\item{chk_name}{chunk name}
}
\value{

}
\description{
Return the depth of chunk that was added into the stack
}
\examples{
GetStackDepth(chunk name)
}
