% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{MultipleChunks}
\alias{MultipleChunks}
\title{MultipleChunks()}
\usage{
MultipleChunks(oper_Time, oper_Name, line)
}
\arguments{
\item{oper_Time}{Operator time (e.g., Look is 550ms)}

\item{oper_Name}{Operator name}

\item{line}{Line number}
}
\value{
Number of chunks in one line
}
\description{
Extract multiple chunks in one vector
}
\examples{
MultipleChunks(operator time, operator name, line number)
}
