% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpm_GUI_1017.R
\name{NCPMcalc}
\alias{NCPMcalc}
\title{NCPMcalc}
\usage{
NCPMcalc(scenario, time_library)
}
\arguments{
\item{scenario}{scenario file}
}
\value{
all outcome
}
\description{
Runs N-CPM model
}
\examples{
NCPMcalc(sce, "Novice")
}
