% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpm_GUI_1017.R
\name{NewOper}
\alias{NewOper}
\title{NewOper}
\usage{
NewOper(op_name, op_time)
}
\arguments{
\item{op_name}{operator name}

\item{op_time}{operator time}
}
\value{

}
\description{
NewOper
}
\examples{
NewOper(a, b)
}
