% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{PushChunk}
\alias{PushChunk}
\title{PushChunk()}
\usage{
PushChunk(
  new_chunk,
  wm_Box,
  time_current,
  oper,
  oper_Time,
  num_of_chunks,
  time_past,
  multiple_chunk_counter,
  chunk_Lifecyle
)
}
\arguments{
\item{new_chunk}{Chunk name}

\item{wm_Box}{A data frame for working memory storage}

\item{time_current}{Current time}

\item{oper}{Operator name}

\item{oper_Time}{Operator time}

\item{num_of_chunks}{Number of chunk in one line}

\item{time_past}{Time when a previous operator was added}

\item{multiple_chunk_counter}{A number to make a difference if there are more than two chunks in one line ???}

\item{chunk_Lifecyle}{A data frame with all chunk's birth and death}
}
\value{
A list of chunk lifecycle and working memory
}
\description{
Push a new chunk
}
\examples{
PuchChunk(new_chunk, wm_Box, time_current, oper, oper_Time, num_of_chunks, time_past, multiple_chunk_counter, chunk_Lifecyle)
}
