% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initiation_data.R
\name{ResetAll}
\alias{ResetAll}
\title{Reset all variables}
\usage{
ResetAll()
}
\value{
Empty wm_Box, chunk_Lifecycle, acc_Time_1, acc_Time_2
}
\description{
Reset all variables
}
\examples{
ResetAll()
}
