% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCT.R
\name{RetrievingOperTime}
\alias{RetrievingOperTime}
\title{RetrievingOperTime()}
\usage{
RetrievingOperTime(oper, k, scenario, skill, oper_set)
}
\arguments{
\item{oper}{operator name}

\item{k}{line number}

\item{scenario}{Scenario in CSV format}

\item{skill}{Novice or Expert}
}
\value{
operator time for an operator
}
\description{
Retrieving the operator time
}
\examples{
RetrievingOperTime(operator name, line number, scenario, skill)
}
