% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCT.R
\name{RetrOpTime_Motor}
\alias{RetrOpTime_Motor}
\title{RetrOpTime_Motor()}
\usage{
RetrOpTime_Motor(oper, k)
}
\arguments{
\item{oper}{operator name}

\item{k}{line number}
}
\value{
error time for a motor operator
}
\description{
Retrieving the operator time (for error)
}
\examples{
RetrOpTime_Motor(operator name, line number)
}
