% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RunMain}
\alias{RunMain}
\title{RunMain()}
\usage{
RunMain(scenario, time_library, skill)
}
\arguments{
\item{scenario}{Scenario in CSV format}

\item{skill}{Novice or Expert}
}
\value{

}
\description{
Run main function
}
\examples{
RunMain(csv file, "Novice")
}
