% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{UpdateChunk}
\alias{UpdateChunk}
\title{UpdateChunk()}
\usage{
UpdateChunk(
  candidate_chunk,
  oper_name,
  time_current,
  wm_Box,
  same_chunk_counter
)
}
\arguments{
\item{candidate_chunk}{Chunk name}

\item{oper_name}{Operator name}

\item{time_current}{Current time}

\item{wm_Box}{A data frame for working memory storage}

\item{same_chunk_counter}{A variable to see if two chunks are same or not}
}
\value{
working memory storage
}
\description{
Update time, rehearsal, activation, and recall prob of old chunks, not the brand new chunk
}
\examples{
UpdateChunk(candidate_chunk, oper_name, time_current, wm_Box, same_chunk_counter)
}
