% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpm_GUI.R, R/ncpm_GUI_0913.R
\name{NCPMcalc}
\alias{NCPMcalc}
\title{NCPMcalc}
\usage{
NCPMcalc(scenario)

NCPMcalc(scenario)
}
\arguments{
\item{scenario}{scenario file}
}
\value{
all outcome

all outcome
}
\description{
Runs N-CPM model

Runs N-CPM model
}
\examples{
NCPMcalc(sce, "Novice")
NCPMcalc(sce, "Novice")
}
