% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpm_GUI_1017.R
\name{NCPMcalc_exp}
\alias{NCPMcalc_exp}
\title{NCPMcalc_expert}
\usage{
NCPMcalc_exp(scenario, time_library)
}
\arguments{
\item{scenario}{scenario file}
}
\value{
all outcome - expert
}
\description{
Runs N-CPM model in an expert's performance
}
\examples{
NCPMcalc_exp(sce, "Expert")
}
