% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpm_GUI_1017.R
\name{NCPMGUI}
\alias{NCPMGUI}
\title{NCPMGUI()}
\usage{
NCPMGUI()
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
GUI

contents for ui
}
\description{
All GUI code
}
\examples{
NCPMGUI()
Shiny Server

no example
}
