% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCT.R
\name{RetrievingOperTime}
\alias{RetrievingOperTime}
\title{RetrievingOperTime()}
\usage{
RetrievingOperTime(oper, k, scenario)
}
\arguments{
\item{oper}{operator name}

\item{k}{line number}
}
\value{
operator time for an operator
}
\description{
Retrieving the operator time
}
\examples{
RetrievingOperTime(operator name, line number)
}
